api/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
pytz==2021.1
sqlparse==0.4.1
djangorestframework